# Summary

This PR changes the array copying process from a for loop to `Array.Copy` in `ImmutableArray.Create`.
This change is expected to improve performance in most cases, except when the array size is small.

# Benchmark

```
| Method    | Size  | Mean         | Error      | StdDev     | Median       | Min          | Max          | Gen0   | Allocated |
|---------- |------ |-------------:|-----------:|-----------:|-------------:|-------------:|-------------:|-------:|----------:|
| Original  | 1     |     2.354 ns |  0.0181 ns |  0.0152 ns |     2.351 ns |     2.334 ns |     2.386 ns | 0.0038 |      32 B |
| PR        | 1     |     3.851 ns |  0.0413 ns |  0.0366 ns |     3.855 ns |     3.776 ns |     3.923 ns | 0.0038 |      32 B |
| Original  | 10    |     5.547 ns |  0.0845 ns |  0.0791 ns |     5.532 ns |     5.431 ns |     5.685 ns | 0.0076 |      64 B |
| PR        | 10    |     4.749 ns |  0.0429 ns |  0.0359 ns |     4.747 ns |     4.706 ns |     4.836 ns | 0.0076 |      64 B |
| Original  | 100   |    41.864 ns |  0.3312 ns |  0.2936 ns |    41.923 ns |    41.391 ns |    42.290 ns | 0.0506 |     424 B |
| PR        | 100   |    15.815 ns |  0.1445 ns |  0.1351 ns |    15.847 ns |    15.612 ns |    16.005 ns | 0.0507 |     424 B |
| Original  | 1000  |   348.828 ns |  8.7576 ns |  9.3705 ns |   344.893 ns |   341.290 ns |   373.834 ns | 0.4804 |    4024 B |
| PR        | 1000  |   144.980 ns |  0.5494 ns |  0.4588 ns |   144.958 ns |   144.272 ns |   145.781 ns | 0.4805 |    4024 B |
| Original  | 10000 | 3,482.065 ns | 27.8280 ns | 26.0303 ns | 3,482.618 ns | 3,446.399 ns | 3,526.556 ns | 4.7491 |   40024 B |
| PR        | 10000 | 1,810.896 ns | 25.1246 ns | 20.9802 ns | 1,808.056 ns | 1,778.418 ns | 1,847.208 ns | 4.7554 |   40024 B |
```

```csharp
    [BenchmarkCategory(Categories.Libraries, Categories.Collections, Categories.GenericCollections)]
    public class ImmutableArrayCreateBenchmark
    {
        private int[] _array;

        [Params(1, 10, 100, 1000, 10000)]
        public int Size { get; set; }

        [GlobalSetup]
        public void GlobalSetup()
        {
            _array = Enumerable.Range(0, Size).ToArray();
        }

        [Benchmark]
        public ImmutableArray<int> Original()
        {
            return ImmutableArray.Create(_array, 0, Size);
        }

        [Benchmark]
        public ImmutableArray<int> PR()
        {
            return ImmutableArray.CreatePR(_array, 0, Size);
        }
    }
```