@xtqqczze 2025-09-19
Is there a benchmark for `ImmutableArray.Create` in https://github.com/dotnet/performance?