@prozolic 2025-09-19
@xtqqczze The benchmark including the case using `ReadOnlySpan.CopyTo`(PRForSpan) is as follows. I don't think this exists in https://github.com/dotnet/performance.

```
BenchmarkDotNet v0.14.1-nightly.20250107.205, Windows 11 (10.0.26100.6584)
11th Gen Intel Core i7-11700 2.50GHz, 1 CPU, 16 logical and 8 physical cores
.NET SDK 10.0.100-rc.1.25451.107
  [Host]     : .NET 10.0.0 (10.0.25.45207), X64 RyuJIT AVX-512F+CD+BW+DQ+VL+VBMI
  Job-QBVWJU : .NET 10.0.0 (42.42.42.42424), X64 RyuJIT AVX-512F+CD+BW+DQ+VL+VBMI

PowerPlanMode=00000000-0000-0000-0000-000000000000  Toolchain=CoreRun  IterationTime=250ms  
MaxIterationCount=20  MinIterationCount=15  WarmupCount=1  

| Method     | Size  | Mean         | Error      | StdDev     | Median       | Min          | Max          | Gen0   | Allocated |
|----------- |------ |-------------:|-----------:|-----------:|-------------:|-------------:|-------------:|-------:|----------:|
| Original   | 1     |     3.271 ns |  0.1850 ns |  0.1900 ns |     3.205 ns |     3.058 ns |     3.831 ns | 0.0038 |      32 B |
| PRForArray | 1     |     4.908 ns |  0.1845 ns |  0.2125 ns |     4.921 ns |     4.590 ns |     5.354 ns | 0.0038 |      32 B |
| PRForSpan  | 1     |     4.599 ns |  0.1736 ns |  0.1929 ns |     4.559 ns |     4.350 ns |     5.012 ns | 0.0038 |      32 B |
| Original   | 10    |     7.224 ns |  0.3928 ns |  0.4366 ns |     7.169 ns |     6.505 ns |     8.157 ns | 0.0076 |      64 B |
| PRForArray | 10    |     5.927 ns |  0.1715 ns |  0.1835 ns |     5.911 ns |     5.574 ns |     6.351 ns | 0.0076 |      64 B |
| PRForSpan  | 10    |     5.807 ns |  0.1536 ns |  0.1707 ns |     5.795 ns |     5.493 ns |     6.177 ns | 0.0076 |      64 B |
| Original   | 100   |    49.718 ns |  1.7005 ns |  1.7463 ns |    49.150 ns |    47.235 ns |    52.983 ns | 0.0506 |     424 B |
| PRForArray | 100   |    19.353 ns |  0.8808 ns |  0.9045 ns |    19.217 ns |    18.588 ns |    21.786 ns | 0.0507 |     424 B |
| PRForSpan  | 100   |    19.493 ns |  0.6797 ns |  0.7827 ns |    19.094 ns |    18.715 ns |    21.188 ns | 0.0506 |     424 B |
| Original   | 1000  |   404.877 ns |  8.7935 ns |  9.7739 ns |   402.753 ns |   386.951 ns |   421.066 ns | 0.4804 |    4024 B |
| PRForArray | 1000  |   172.957 ns |  5.4354 ns |  6.2594 ns |   171.944 ns |   165.877 ns |   186.308 ns | 0.4808 |    4024 B |
| PRForSpan  | 1000  |   169.276 ns |  5.4341 ns |  5.8144 ns |   168.339 ns |   160.069 ns |   182.131 ns | 0.4807 |    4024 B |
| Original   | 10000 | 3,820.823 ns | 57.4370 ns | 53.7266 ns | 3,816.897 ns | 3,763.743 ns | 3,944.127 ns | 4.7551 |   40024 B |
| PRForArray | 10000 | 2,022.101 ns | 23.5543 ns | 18.3897 ns | 2,018.220 ns | 1,988.932 ns | 2,049.105 ns | 4.7610 |   40024 B |
| PRForSpan  | 10000 | 2,046.557 ns | 49.5834 ns | 55.1119 ns | 2,032.679 ns | 1,995.516 ns | 2,178.990 ns | 4.7617 |   40024 B |
```