@xtqqczze 2025-09-20
> The benchmark including the case using `ReadOnlySpan.CopyTo`(PRForSpan) is as follows.

So there's no difference in throughput. Is there any difference is code size? You will need something like BDN `DisassemblyDiagnoser`